[libs-update] Extracted renderPage in a separate helper to avoid break the jest tests.
