+++ assets/pages/bundles-landing/bundlesLanding.jsx
@@ -9,7 +9,8 @@ import SimpleHeader from 'components/headers/simpleHeader/simpleHeader';
 import LinksFooter from 'components/footers/linksFooter/linksFooter';
 
 import { trackOphan } from 'helpers/abtest';
-import { init as pageInit, renderPage } from 'helpers/page/page';
+import { init as pageInit } from 'helpers/page/page';
+import { renderPage } from 'helpers/render';
 
 import Introduction from './components/Introduction';
 import Bundles from './components/Bundles';
