+++ assets/pages/oneoff-contributions/oneoffContributions.jsx
@@ -22,7 +22,8 @@ import { termsLinks } from 'helpers/internationalisation/legal';
 import * as user from 'helpers/user/user';
 import { getQueryParameter } from 'helpers/url';
 import { parse as parseContrib } from 'helpers/contributions';
-import { init as pageInit, renderPage } from 'helpers/page/page';
+import { init as pageInit } from 'helpers/page/page';
+import { renderPage } from 'helpers/render';
 
 import OneoffContributionsPayment from './components/oneoffContributionsPayment';
 import FormFields from './components/formFields';
